requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiohttp==3.9.1

# LangChain and LLM Integration
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langchain-core==0.1.10
openai==1.3.7
anthropic==0.7.8

# Database
sqlalchemy==2.0.23
alembic==1.13.1
sqlite3

# Configuration and Utilities
pyyaml==6.0.1
python-dotenv==1.0.0

# Async and Concurrency
asyncio
aiofiles==23.2.1

# HTTP and API
requests==2.31.0
httpx==0.25.2

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Date and Time
python-dateutil==2.8.2

# Logging and Monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: Machine Learning (for advanced features)
scikit-learn==1.3.2

# Optional: Natural Language Processing
nltk==3.8.1
spacy==3.7.2

# Optional: Web Scraping (for deal finder)
beautifulsoup4==4.12.2
selenium==4.15.2

# Optional: Calendar Integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Optional: Health and Fitness APIs
fitbit==0.3.2
stravalib==1.10.0

# Optional: Shopping APIs
ebay-api==0.1.0
amazon-paapi==5.0.0

# Optional: Weather (for outdoor activities)
requests-oauthlib==1.3.1
